requirements.txt
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
dataclasses==0.6
dj-database-url==3.1.2
Django==6.0.4
django-stubs==6.0.2
django-stubs-ext==6.0.2
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
djongo==1.2.31
dnspython==2.8.0
flet==0.84.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
msgpack==1.1.2
oauthlib==3.3.1
psycopg2-binary==2.9.11
pymongo==4.16.0
python-dotenv==1.2.2
pytz==2026.1.post1
repath==0.9.0
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
urllib3==2.6.3
